requirements.txt
aioredis==1.3.1
alpaca-trade-api==0.51.0
asgi-redis==1.4.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
dj-database-url==0.5.0
Django==3.1.5
django-autoslug==1.9.8
django-crispy-forms==1.9.2
django-dash==0.6.1
django-heroku==0.3.1
django-nine==0.2.4
django-tinymce==3.2.0
easy-thumbnails==2.7.1
feedparser==6.0.2
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
msgpack==1.0.2
msgpack-python==0.5.6
numpy==1.19.2
pandas==1.1.3
pif==0.8.2
Pillow==8.1.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.5
redis==2.10.6
requests==2.25.1
service-identity==18.1.0
sgmllib3k==1.0.0
six==1.15.0
sqlparse==0.4.1
transliterate==1.10.2
Twisted==20.3.0
txaio==20.12.1
Unidecode==1.2.0
urllib3==1.26.2
vishap==0.1.5
websocket-client==0.57.0
websockets==8.1
whitenoise==5.2.0
zope.interface==5.2.0
